board monthy feature and add icons

+++ lib/features/board/widgets/item_mood_percent_detail.dart
@@ -5,7 +5,7 @@ import 'package:flutter/src/widgets/placeholder.dart';
 
 class ItemMoodPercentDetail extends StatelessWidget {
   final Color color;
-  final double percent;
+  final int percent;
 
   const ItemMoodPercentDetail({
     super.key,
